doc: synthetic Integrations
