+++ e2e/commonwithcustominstall/files/deploy.yaml
